+++ NEWS.md
@@ -15,6 +15,7 @@
 - `eeg_combine.tfr_average()` added to handle pre-averaged `eeg_tfr` objects
 - `compute_tfr()` now allows non-constant number of cycles
 - `compute_tfr()` now uses a different scaling factor, so raw units should now be microvolts-squared. 
+- added `import_erplab()` function
 
 ### Internal changes / bug fixes
 - `plot_tfr()` now always drops NA/NaN values and averages appropriately over electrodes and conditions.
@@ -40,6 +41,7 @@
 - `convert_tfr()` now properly returns converted data
 - `import_raw()` fix for Brain Vision Analyzer files with date fields in the markers
 - added `version` field to most objects
+
 
 # eegUtils 0.5.0